maybe i fixed it

+++ WhoAmIBotNode/Classes/NodeGroup.cs
@@ -6,8 +6,8 @@
         public string LangKey { get; set; }
         public string Name { get; set; }
         public bool CancelgameAdmin { get; set; } = true;
-        public long JoinTimeout { get; set; } = 10;
-        public long GameTimeout { get; set; } = 1440;
+        public int JoinTimeout { get; set; } = 10;
+        public int GameTimeout { get; set; } = 1440;
         public NodeGroup(long id)
         {
             Id = id;
